Time: 1 ms (98.36%), Space: 41.2 MB (28.94%) - LeetHub

+++ 0441-arranging-coins/0441-arranging-coins.java
@@ -4,7 +4,7 @@ class Solution {
         long max=n;
         long ans=0;
         while(min<=max){
-            long mid = (long)(min+max)/2;
+            long mid = (min+max)/2;
             if(isPossible(n,mid)){
                 ans=mid;
                 min=mid+1;
